+++ packages/apollo-link-state/src/index.ts
@@ -20,7 +20,7 @@ const capitalizeFirstLetter = str => str.charAt(0).toUpperCase() + str.slice(1);
 
 export type ClientStateConfig = {
   cache?: ApolloCache<any>;
-  resolvers: any;
+  resolvers: any | (() => any);
   defaults?: any;
   typeDefs?: string | string[];
   fragmentMatcher?: FragmentMatcher;
@@ -29,13 +29,7 @@ export type ClientStateConfig = {
 export const withClientState = (
   clientStateConfig: ClientStateConfig = { resolvers: {}, defaults: {} },
 ) => {
-  const {
-    resolvers,
-    defaults,
-    cache,
-    typeDefs,
-    fragmentMatcher,
-  } = clientStateConfig;
+  const { defaults, cache, typeDefs, fragmentMatcher } = clientStateConfig;
   if (cache && defaults) {
     cache.writeData({ data: defaults });
   }
@@ -67,6 +61,10 @@ export const withClientState = (
 
       if (!isClient) return forward(operation);
 
+      const resolvers =
+        typeof clientStateConfig.resolvers === 'function'
+          ? clientStateConfig.resolvers()
+          : clientStateConfig.resolvers;
       const server = removeClientSetsFromDocument(operation.query);
       const { query } = operation;
       const type =