+++ packages/apollo-link-state/src/utils.ts
@@ -1,9 +1,6 @@
 import { DocumentNode, DirectiveNode } from 'graphql';
 
-import {
-  checkDocument,
-  removeDirectivesFromDocument,
-} from 'apollo-utilities';
+import { checkDocument, removeDirectivesFromDocument } from 'apollo-utilities';
 
 const connectionRemoveConfig = {
   test: (directive: DirectiveNode) => directive.name.value === 'client',